Update page bitmap helpers to use 32 bit addresses

+++ tile/kernel/memory.h
@@ -83,8 +83,8 @@ struct memory_map {
 */
 struct memory_bitmap {
   uint32_t* data;
-  uint64_t size;
-  uint64_t offset;
+  uint32_t size;
+  uint32_t offset;
   struct memory_bitmap* next;
 };
 
@@ -115,9 +115,9 @@ void memory_map_insert_block(struct memory_map_group* group, int pos, uint64_t b
 void memory_map_add_block(struct memory_map_group* group, uint64_t begin, uint64_t size);
 int memory_map_split_block(struct memory_map_group* group, uint64_t begin);
 
-size_t bitmap_index(const struct memory_bitmap* bitmap, uint64_t addr);
-size_t bitmap_index_index(const struct memory_bitmap* bitmap, uint64_t addr);
-uint64_t bitmap_to_addr(const struct memory_bitmap* bitmap, uint64_t i, uint64_t j);
+size_t bitmap_index(const struct memory_bitmap* bitmap, uint32_t addr);
+size_t bitmap_index_index(const struct memory_bitmap* bitmap, uint32_t addr);
+uint32_t bitmap_to_addr(const struct memory_bitmap* bitmap, size_t i, size_t j);
 void bitmap_insert(struct memory_bitmap* bitmap, uint32_t addr, uint32_t size);
 void bitmap_clear(struct memory_bitmap* bitmap, uint32_t addr);
